add devlog for babycastles

+++ mov2mp4.sh
@@ -1,3 +1,6 @@
+# usage: 
+# mov2mp4.sh ./documentation/2022-12-03/*.MOV
+
 # Go through input files and convert all to mp4
 # delete originals (be careful!)
 
@@ -5,16 +8,16 @@
 
 for f in "$@"
 do
-  if [[ $f == *.MOV ]]
+  if [[ "$f" == *.MOV ]] && [[ -f "$f" ]]
   then
     of="${f%.MOV}.mp4"
     cmd="ffmpeg -y -i $f -vcodec h264 $of"
-    echo cmd
+    echo $cmd
     $cmd
 
     cmd="rm $f"
     rm $f
-    echo cmd
+    echo $cmd
     $cmd
   fi
 done
